separate connection test to multiple files
